+++ client/java/src/test/java/io/openlineage/client/OpenLineageTest.java
@@ -15,8 +15,10 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.openlineage.client.OpenLineage.DataQualityMetricsInputDatasetFacet;
 import io.openlineage.client.OpenLineage.DataQualityMetricsInputDatasetFacetColumnMetricsAdditional;
+import io.openlineage.client.OpenLineage.DatasetEvent;
 import io.openlineage.client.OpenLineage.InputDataset;
 import io.openlineage.client.OpenLineage.Job;
+import io.openlineage.client.OpenLineage.JobEvent;
 import io.openlineage.client.OpenLineage.JobFacets;
 import io.openlineage.client.OpenLineage.NominalTimeRunFacet;
 import io.openlineage.client.OpenLineage.OutputDataset;
@@ -27,6 +29,7 @@ import java.net.URI;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
@@ -46,7 +49,7 @@ class OpenLineageTest {
   }
 
   @Test
-  void jsonSerialization() throws JsonProcessingException {
+  void jsonRunEventSerialization() throws JsonProcessingException {
     ZonedDateTime now = ZonedDateTime.now(ZoneId.of("UTC"));
 
     URI producer = URI.create("producer");
@@ -86,6 +89,53 @@ class OpenLineageTest {
     assertEquals("output", outputDataset.getName());
 
     assertEquals(roundTrip(json), roundTrip(mapper.writeValueAsString(read)));
+  }
+
+  @Test
+  void jsonDatasetEventSerialization() throws JsonProcessingException {
+    ZonedDateTime now = ZonedDateTime.now(ZoneId.of("UTC"));
+    URI producer = URI.create("producer");
+    OpenLineage ol = new OpenLineage(producer);
+
+    DatasetEvent datasetEvent =
+        ol.newDatasetEventBuilder()
+            .eventTime(now)
+            .dataset(ol.newStaticDataset("ns", "ds", null))
+            .build();
+
+    String json = mapper.writeValueAsString(datasetEvent);
+    DatasetEvent read = mapper.readValue(json, DatasetEvent.class);
+
+    assertEquals("ns", read.getDataset().getNamespace());
+    assertEquals("ds", read.getDataset().getName());
+    assertEquals(now, read.getEventTime());
+  }
+
+  @Test
+  void jsonJobEventSerialization() throws JsonProcessingException {
+    ZonedDateTime now = ZonedDateTime.now(ZoneId.of("UTC"));
+    URI producer = URI.create("producer");
+    OpenLineage ol = new OpenLineage(producer);
+    JobEvent jobEvent =
+        ol.newJobEventBuilder()
+            .eventTime(now)
+            .inputs(Collections.singletonList(ol.newInputDataset("ins", "input", null, null)))
+            .outputs(Collections.singletonList(ol.newOutputDataset("ous", "output", null, null)))
+            .job(ol.newJob("jns", "jname", null))
+            .build();
+
+    String json = mapper.writeValueAsString(jobEvent);
+    JobEvent read = mapper.readValue(json, JobEvent.class);
+
+    assertEquals("ins", read.getInputs().get(0).getNamespace());
+    assertEquals("input", read.getInputs().get(0).getName());
+
+    assertEquals("ous", read.getOutputs().get(0).getNamespace());
+    assertEquals("output", read.getOutputs().get(0).getName());
+
+    assertEquals("jns", read.getJob().getNamespace());
+    assertEquals("jname", read.getJob().getName());
+    assertEquals(now, read.getEventTime());
   }
 
   String roundTrip(String value) throws JsonMappingException, JsonProcessingException {