+++ client/java/src/test/java/io/openlineage/client/transports/KafkaTransportTest.java
@@ -5,7 +5,7 @@
 
 package io.openlineage.client.transports;
 
-import static io.openlineage.client.Events.event;
+import static io.openlineage.client.Events.runEvent;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
@@ -39,7 +39,7 @@ class KafkaTransportTest {
 
     when(producer.send(any(ProducerRecord.class))).thenReturn(mock(Future.class));
 
-    client.emit(event());
+    client.emit(runEvent());
 
     ArgumentCaptor<ProducerRecord<String, String>> captor =
         ArgumentCaptor.forClass(ProducerRecord.class);